+++ Action/Action.php
@@ -4,7 +4,7 @@ namespace LoremIpsum\ActionLoggerBundle\Action;
 
 use App\Entity\User;
 use LoremIpsum\ActionLoggerBundle\Entity\LogAction;
-use LoremIpsum\RouteGeneratorBundle\RouteGeneratorInterface;
+use LoremIpsum\RouteGeneratorBundle\Model\RouteGeneratorInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 
 abstract class Action implements ActionInterface