+++ Action/UnknownAction.php
@@ -3,7 +3,7 @@
 namespace LoremIpsum\ActionLoggerBundle\Action;
 
 use LoremIpsum\ActionLoggerBundle\Entity\LogAction;
-use LoremIpsum\RouteGeneratorBundle\RouteGeneratorInterface;
+use LoremIpsum\RouteGeneratorBundle\Model\RouteGeneratorInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 
 final class UnknownAction extends Action